[010-Help엔티티들-제작] PR 반영
