--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/common/exception/DomainException.java
@@ -1,10 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.common.exception;
-
-
-
-public class DomainException extends RuntimeException{
-
-    public DomainException(String errorMessage) {
-        super(errorMessage);
-    }
-}
