--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/common/exception/WebException.java
@@ -1,10 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.common.exception;
-
-
-
-public class WebException extends RuntimeException{
-
-    public WebException(String errorMessage){
-        super(errorMessage);
-    }
-}
