+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/infra/exceptions/JPAException.java
@@ -1,0 +1,9 @@
+package com.example.checkinrequestMS.HelpAPI.infra.exceptions;
+
+import com.example.checkinrequestMS.common.exception.types.InfraException;
+
+public class JPAException extends InfraException {
+    public JPAException(JPAErrorCode e) {
+        super(e.getDetail());
+    }
+}
