--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/PlaceAPI/common/exception/response/ExceptionResponse.java
@@ -1,15 +1,0 @@
-package com.example.checkinrequestMS.PlaceAPI.common.exception.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@ToString
-@AllArgsConstructor
-public class ExceptionResponse {
-    private HttpStatus status;
-    private String message;
-}
-
