+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/common/exception/types/DomainException.java
@@ -1,9 +1,8 @@
-package com.example.checkinrequestMS.PlaceAPI.common.exception;
+package com.example.checkinrequestMS.common.exception.types;
 
 
 
 public class DomainException extends RuntimeException{
-
     public DomainException(String errorMessage) {
         super(errorMessage);
     }
