+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/common/exception/types/InfraException.java
@@ -1,0 +1,8 @@
+package com.example.checkinrequestMS.common.exception.types;
+
+public class InfraException extends RuntimeException{
+
+    public InfraException(String errorMessage) {
+        super(errorMessage);
+    }
+}
